Projects/Alpha Analytics/requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
scipy>=1.10
scikit-learn>=1.3
yfinance>=0.2.40
requests>=2.31
python-dotenv>=1.0
fredapi>=0.5
jupyter>=1.0
